Adding code files
BubbleSort.pde file has the "drawings" part of the code, which is quite simple. Meanwhile, the "logics" part of the code is inside Mem.pde file (Mem for Memory). As for the BubbleSort.mp4 file, there is a working example of the code's execution, IF whoever is reading would like to watch it.